Make build failures easier to detect
If /usr/lib/live/build/binary_hdd fails to unmount the temporary VFAT
filesystem then the build halts, but 'lh build' did not exit with an
error. The use of the wildcard in the Makefile means that we did not
notice the lack of the output file binary.img.

Hardcode the list of expected output files for each build type, and fail
the build if they are missing.