poclbm-wrapper: make more robust
If the setup portion of the script failed, the script would exit and the
user wouldn't see the error message indicating why. Move the setup
portion within the while loop.

In addition, the exit status of any failing command within the subshell
propogated to the subshell as a whole, aborting the script. The only
safe way to ignore the exit status of the subshell is to set +e
immediately before it.